+++ source/scener/graphics/vertex_declaration.hpp
@@ -18,19 +18,19 @@ namespace scener::graphics
         /// Initializes a new instance of the VertexDeclaration class.
         /// \param stride The number of bytes per element.
         /// \param elements vertex elements.
-        vertex_declaration(std::size_t stride, const std::vector<vertex_element>& elements) noexcept;
+        vertex_declaration(std::uint32_t stride, const std::vector<vertex_element>& elements) noexcept;
 
     public:
         /// Gets the number of bytes from one vertex to the next.
         /// \returns the stride (in bytes).
-        std::size_t vertex_stride() const noexcept;
+        std::uint32_t vertex_stride() const noexcept;
 
         /// Gets the vertex shader declaration.
         /// \returns the vertex elements that make up the vertex shader declaration.
         const std::vector<vertex_element>& vertex_elements() const noexcept;
 
     private:
-        std::size_t                 _vertex_stride;
+        std::uint32_t               _vertex_stride;
         std::vector<vertex_element> _vertex_elements;
     };
 }
